api D E S C A N S A N T E

+++ Chat2/Controllers/Messages/MessagesController.cs
@@ -1,4 +1,6 @@
-﻿using Chat2.model;
+﻿using AutoMapper;
+using Chat2.Controllers.Base;
+using Chat2.model;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
@@ -8,24 +10,26 @@ namespace Chat2.Controllers;
 
 [ApiController]
 [Route("/rooms/{roomId}/[controller]")]
-public class MessagesController : Controller
+public class MessagesController : BaseController
 {
     private readonly IConfiguration _config;
     private readonly Context _context;
-    private UserManager<User> UserManager;
+    private IMapper _mapper;
 
-    public MessagesController(Context context, IConfiguration config, UserManager<User> userManager)
+    public MessagesController(Context context, IConfiguration config, UserManager<User> userManager, IMapper mapper) :
+        base(userManager)
     {
         _context = context;
         _config = config;
-        UserManager = userManager;
+        _mapper = mapper;
     }
 
     [HttpPost]
     [Authorize]
+    [Route("/me/message")]
     public async Task<IActionResult> CreateMessage(string roomId, CreateMessageDto createMessageDto)
     {
-        var author = await UserManager.GetUserAsync(User);
+        var author = await GetUserOrFailAsync();
         var room = await _context.Rooms.FindAsync(roomId);
         var message = new Message
         {
@@ -36,21 +40,7 @@ public class MessagesController : Controller
         };
         await _context.Messages.AddAsync(message);
         await _context.SaveChangesAsync();
-        var messageDto = new MessageDto
-        {
-            Author = new UserDetailsDto
-            {
-                Id = author.Id,
-                UserName = author.UserName
-            },
-            Body = message.Body,
-            CreatedAt = message.CreatedAt,
-            Room = new RoomDetailsDto
-            {
-                Id = room.Id,
-                Name = room.Name
-            }
-        };
+        var messageDto = _mapper.Map<MessageDetailsDto>(message);
         return Ok(messageDto);
     }
 
@@ -59,32 +49,51 @@ public class MessagesController : Controller
     public IActionResult GetMessage(string messageId)
     {
         var message = _context.Messages.Include("Author").Include("Room").First(m => m.Id == messageId);
-        var getmessage = new MessageDetailsDto
-        {
-            Body = message.Body,
-            RoomName = message.Room.Name,
-            AuthorName = message.Author.UserName
-        };
+        var getmessage = _mapper.Map<MessageDetailsDto>(message);
         return Ok(getmessage);
     }
 
-
     [HttpPatch]
-    public IActionResult EditMessage(EditMessageDto editMessageDto)
+    [Authorize]
+    [Route("/me/{messageId}")]
+    public async Task<IActionResult> EditMessage(string messageId, EditMessageDto editMessageDto)
     {
-        var message = _context.Messages.First(m => m.Id == editMessageDto.Id);
-        message.Body = editMessageDto.NewBody;
-        message.Edited = true;
-        _context.SaveChanges();
-        return Ok();
+        var user = await GetUserOrFailAsync();
+        var message = await _context.Messages.FindAsync(messageId);
+        if (message != null)
+        {
+            if (message.Author == user)
+            {
+                message.Body = editMessageDto.NewBody;
+                message.Edited = true;
+                await _context.SaveChangesAsync();
+                return Ok();
+            }
+
+            return Unauthorized();
+        }
+
+        return NotFound();
     }
 
     [HttpDelete]
-    [Route("{messageId}")]
-    public void DeleteMessage(string messageId)
+    [Authorize]
+    [Route("/me/{messageId}")]
+    public async Task<IActionResult> DeleteMessage(string messageId)
     {
-        var message = _context.Messages.First(m => m.Id == messageId);
-        _context.Messages.Remove(message);
-        _context.SaveChanges();
+        var user = await GetUserOrFailAsync();
+        var message = await _context.Messages.FindAsync(messageId);
+        if (message != null)
+        {
+            if (message.Author == user)
+            {
+                _context.Messages.Remove(message);
+                await _context.SaveChangesAsync();
+            }
+
+            return Unauthorized();
+        }
+
+        return NotFound();
     }
 }
